pyproject.toml
[project]
name = "exoplanet-detection-api"
version = "0.1.0"
description = "NASA Space Apps - Exoplanet Detection API"
authors = [
    {name = "ECI-Centauri"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=25.9.0,<26.0.0)",
    "pylint (>=3.3.8,<4.0.0)"
]

[tool.poetry]
package-mode = false
name = "exoplanet-detection-api"
version = "0.1.0"
description = "NASA Space Apps - Exoplanet Detection API"
authors = ["ECI Centauri <jesus.pinzon-v@mail.escuelaing.edu.co>"]
readme = "README.md"
packages = [{include = "app"}]
